requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
dominate==2.6.0
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-DotEnv==0.1.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
SQLAlchemy==1.4.27
urllib3==1.26.7
visitor==0.1.3
Werkzeug==2.0.2
WTForms==3.0.0
gunicorn==20.1.0
